Разбивка reducer.js на data.js и state.js
