+++ src/reducer/data/data.js
@@ -1,0 +1,71 @@
+import {extend} from "../../utils/common.js";
+import {adapterMovies, adapterMovie} from "../../utils/movie.js";
+
+const promoMovie = {
+  title: `The Grand Budapest Hotel`,
+  genre: `Drama`,
+  date: `2014`,
+  poster: `img/the-grand-budapest-hotel-poster.jpg`,
+  background: `img/bg-the-grand-budapest-hotel.jpg`,
+  descriptions: [
+    `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
+    `Gustave prides himself on providing first-class service to the hotel's guests, including satisfying the sexual needs of the many elderly women who stay there. When one of Gustave's lovers dies mysteriously, Gustave finds himself the recipient of a priceless painting and the chief suspect in her murder.`,
+  ],
+  score: 6.6,
+  rating: `240`,
+  director: `Wes Andreson`,
+  starring: `Bill Murray, Edward Norton, Jude Law, Willem Dafoe and other`,
+};
+
+const initialState = {
+  movies: [],
+  promoMovie: promoMovie,
+};
+
+const ActionType = {
+  LOAD_MOVIES: `LOAD_MOVIES`,
+  LOAD_PROMO_MOVIE: `LOAD_PROMO_MOVIE`,
+};
+
+const ActionCreator = {
+  loadMovies: (movies) => ({
+    type: ActionType.LOAD_MOVIES,
+    payload: adapterMovies(movies),
+  }),
+  loadPromoMovies: (movie) => ({
+    type: ActionType.LOAD_PROMO_MOVIE,
+    payload: adapterMovie(movie),
+  }),
+};
+
+const DataOperation = {
+  loadMovies: () => (dispatch, getState, api) => {
+    return api.get(`/films`)
+      .then((response) => {
+        dispatch(ActionCreator.loadMovies(response.data));
+      });
+  },
+  loadPromoMovies: () => (dispatch, getState, api) => {
+    return api.get(`/films/promo`)
+      .then((response) => {
+        dispatch(ActionCreator.loadPromoMovies(response.data));
+      });
+  },
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.LOAD_MOVIES:
+      return extend(state, {
+        movies: action.payload,
+      });
+    case ActionType.LOAD_PROMO_MOVIE:
+      return extend(state, {
+        promoMovie: action.payload,
+      });
+  }
+
+  return state;
+};
+
+export {reducer, DataOperation, ActionType, ActionCreator};
