+++ src/reducer/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  DATA: `DATA`,
+  STATE: `STATE`,
+};
+
+export default NameSpace;
