+++ src/reducer/reducer.js
@@ -1,0 +1,10 @@
+import {combineReducers} from "redux";
+import {reducer as data} from "./data/data.js";
+import {reducer as state} from "./state/state.js";
+import NameSpace from "./name-space.js";
+
+
+export default combineReducers({
+  [NameSpace.DATA]: data,
+  [NameSpace.STATE]: state,
+});
