+++ src/utils/common.js
@@ -1,0 +1,3 @@
+const extend = (a, b) => {
+  return Object.assign({}, a, b);
+};
